backend/requirements.txt
absl-py==1.2.0
aniso8601==9.0.1
astunparse==1.6.3
boto3==1.14.56
botocore==1.17.63
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
Cython==0.29.32
docutils==0.15.2
faiss-cpu==1.7.2
Flask==2.2.2
Flask-RESTful==0.3.9
flask-cors
fonttools==4.37.1
future==0.18.2
gast==0.3.3
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==2.10.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.18.5
oauthlib==3.2.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
protobuf==3.19.5
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2019.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.3.7
scipy==1.4.1
six==1.16.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==2.0.1
torch==1.6.0
torchvision==0.7.0
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.25.11
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
gunicorn